M2 S4: serialize answers, and the round trip is a fixed point
`stateToMarkdown.ts` (626 lines) transcribed as `mt_md::serialize`, on top of
muya's own `stringWidth.ts` (60 lines) rather than `unicode-width` — the
fifteen inlined East-Asian ranges are a hand-written list and reproducing the
list is the point, because column width is what `_serializeTable`'s padding
arithmetic is made of.

D6's second ratchet wakes. `crates/mt-md/tests/round_trip.rs` states this
crate's contract over 1346 inputs — 11 corpus files, 11 round-trip fixtures and
all 1324 spec examples:

  parse(serialize(parse(s))) == parse(s)   1324 of 1346, 22 enumerated
  serialize(parse(s))        == s           998 of 1346, 348 enumerated

The fixed point holds over all 22 whole documents with no exceptions, which is
§9's exit-gate wording exactly. Both exception sets are ratchets rather than
reports: a listed input that starts round-tripping fails the build.

And the measurement neither clause asked for: **the port's serialization of all
1346 inputs is byte-identical to `ExportMarkdown.generate`'s**, `5mb.md`,
`cjk.md` and `rtl.md` included. That is what makes both lists readable as facts
about the reference serializer rather than as this crate's losses.

Two things §10 owed this stage, paid:

- `meta.fenceLength` no longer saturates at 255. `fence_len` is `Option<u32>`;
  a 300-backtick fence round-trips, as it does in muya.
- `- [ ] \ntext` folds into the task item. All thirteen of S3's probes agree
  now; the six that already agreed are unmoved.

`serialize` is infallible, decided by `stateToMarkdown.spec.ts` rather than by
the argument: it asks what happens to a document `MarkdownToState` cannot
produce, twice, and both times the answer is degraded output rather than an
error.

`mt-cli::read_markdown` moved into `mt_md::normalize_source` so the round-trip
gate cannot become a fourth spelling of the CRLF rule; `parse` still keeps
carriage returns, because muya's does.

`PENDING` 121 → 42. Nine of the 42 are the finding: `marked` re-lexes a list
item's content line by line, so an empty bullet or a non-1 ordered marker
starts a list inside an item where CommonMark keeps the paragraph open. Owed to
S6 in §10 with its reproducers, because no harness's input set reaches it.

Gates: `xtask diff --require-ts` 22/22, `xtask blocks` 1344/1344 with 4,985
leaves, `xtask divergences` 3 inline entries ok over 27, `xtask conformance`
still skips 1,324. `render_to_static_html` still returns Err(Unimplemented).
773 passed, 0 failed.

+++ crates/mt-doc/src/block.rs
@@ -84,7 +84,25 @@ pub enum Block {
         /// — a fence opened with ```` ```` ```` (four backticks) must be
         /// re-emitted with four, not normalised to three, or a fence
         /// containing a three-backtick sequence silently breaks.
-        fence_len: Option<u8>,
+        ///
+        /// # Widened from `u8` at M2 S4, and the reason is the constraint above
+        ///
+        /// M0 wrote this as `Option<u8>` and M2.md §10's "Owed by S1" recorded
+        /// what that costs: ``"`".repeat(300)`` opens a 300-backtick fence,
+        /// `marked` carries a JavaScript number, and a `u8` saturates at 255 —
+        /// so the serializer would re-emit 255 backticks and the block would no
+        /// longer contain its own body. That is precisely the round-trip loss
+        /// constraint 3 exists to prevent, in the one shape M0's type could not
+        /// express.
+        ///
+        /// It is `u32` rather than `usize` to match
+        /// [`Block::OrderList::start`], which is the other field carrying a
+        /// JavaScript number, and because a fence longer than four billion
+        /// characters is a document nobody has. No fixture in
+        /// `cargo xtask blocks`' 1344 reaches even 256, which is why this
+        /// waited for the stage that would have lost the data rather than
+        /// being fixed where it was found.
+        fence_len: Option<u32>,
 
         text: Text,
     },
@@ -531,7 +549,8 @@ pub enum BlockMeta {
     CodeBlock {
         kind: CodeKind,
         info: String,
-        fence_len: Option<u8>,
+        /// See [`Block::CodeBlock::fence_len`] — widened from `u8` at M2 S4.
+        fence_len: Option<u32>,
     },
     MathBlock {
         style: MathStyle,
